dashboard/requirements.txt
fastapi
uvicorn
websocket-client
websockets
asyncio
python-dotenv
